+++ Example/Segment-ComScore/SEGViewController.m
@@ -7,6 +7,7 @@
 //
 
 #import "SEGViewController.h"
+#import <Analytics/SEGAnalytics.h>
 
 
 @interface SEGViewController ()